// diagram-editor/frontend/types/diagram.ts
export type NextOperation =
  | string
  | {
      builtin: BuiltinTarget;
      [k: string]: unknown;
    }
  | (NamespacedOperation & {
      [k: string]: unknown;
    });
export type BuiltinTarget = 'terminate' | 'dispose' | 'cancel';
export type DiagramOperation =
  | (NodeSchema & {
      type: 'node';
      [k: string]: unknown;
    })
  | (SectionSchema & {
      type: 'section';
      [k: string]: unknown;
    })
  | (ScopeSchema & {
      type: 'scope';
      [k: string]: unknown;
    })
  | (StreamOutSchema & {
      type: 'stream_out';
      [k: string]: unknown;
    })
  | (ForkCloneSchema & {
      type: 'fork_clone';
      [k: string]: unknown;
    })
  | (UnzipSchema & {
      type: 'unzip';
      [k: string]: unknown;
    })
  | (ForkResultSchema & {
      type: 'fork_result';
      [k: string]: unknown;
    })
  | (SplitSchema & {
      type: 'split';
      [k: string]: unknown;
    })
  | (JoinSchema & {
      type: 'join';
      [k: string]: unknown;
    })
  | (SerializedJoinSchema & {
      type: 'serialized_join';
      [k: string]: unknown;
    })
  | (TransformSchema & {
      type: 'transform';
      [k: string]: unknown;
    })
  | (BufferSchema & {
      type: 'buffer';
      [k: string]: unknown;
    })
  | (BufferAccessSchema & {
      type: 'buffer_access';
      [k: string]: unknown;
    })
  | (ListenSchema & {
      type: 'listen';
      [k: string]: unknown;
    });
export type SectionSchema = (
  | {
      builder: string;
      [k: string]: unknown;
    }
  | {
      template: string;
      [k: string]: unknown;
    }
) & {
  config?: {
    [k: string]: unknown;
  };
  connect?: {
    [k: string]: NextOperation;
  };
  [k: string]: unknown;
};
export type BufferSelection =
  | {
      [k: string]: NextOperation;
    }
  | NextOperation[];
// ...
export interface NamespacedOperation {
  [k: string]: string;
}
export interface NodeSchema {
  builder: string;
  config?: unknown;
  next: NextOperation;
  stream_out?: {
    [k: string]: NextOperation;
  };
  [k: string]: unknown;
}
/**
 * The schema to define a scope within a diagram.
 */
export interface ScopeSchema {
  next: NextOperation;
  /**
   * To simplify diagram definitions, the diagram workflow builder will
   *  sometimes insert implicit operations into the workflow, such as implicit
   *  serializing and deserializing. These implicit operations may be fallible.
   *
   *  This field indicates how a failed implicit operation should be handled.
   *  If left unspecified, an implicit error will cause the entire workflow to
   *  be cancelled.
   */
  on_implicit_error?: NextOperation | null;
  /**
   * Operations that exist inside this scope.
   */
  ops: {
    [k: string]: DiagramOperation;
  };
  settings?: ScopeSettings;
  start: NextOperation;
  /**
   * Where to connect streams that are coming out of this scope.
   */
  stream_out?: {
    [k: string]: NextOperation;
  };
  [k: string]: unknown;
}
/**
 * Settings which determine how the top-level scope of the workflow behaves.
 */
export interface ScopeSettings {
  /**
   * Should we prevent the scope from being interrupted (e.g. cancelled)?
   *  False by default, meaning by default scopes can be cancelled or
   *  interrupted.
   */
  uninterruptible: boolean;
  [k: string]: unknown;
}
export interface StreamOutSchema {
  /**
   * The name of the stream exiting the workflow or scope.
   */
  name: string;
  [k: string]: unknown;
}
export interface ForkCloneSchema {
  next: NextOperation[];
  [k: string]: unknown;
}
export interface UnzipSchema {
  next: NextOperation[];
  [k: string]: unknown;
}
export interface ForkResultSchema {
  err: NextOperation;
  ok: NextOperation;
  [k: string]: unknown;
}
export interface SplitSchema {
  keyed?: {
    [k: string]: NextOperation;
  };
  remaining?: NextOperation | null;
  sequential?: NextOperation[];
  [k: string]: unknown;
}
export interface JoinSchema {
  buffers: BufferSelection;
  next: NextOperation;
  [k: string]: unknown;
}
export interface SerializedJoinSchema {
  buffers: BufferSelection;
  next: NextOperation;
  [k: string]: unknown;
}
export interface TransformSchema {
  cel: string;
  next: NextOperation;
  /**
   * Specify what happens if an error occurs during the transformation. If
   *  you specify a target for on_error, then an error message will be sent to
   *  that target. You can set this to `{ "builtin": "dispose" }` to simply
   *  ignore errors.
   *
   *  If left unspecified, a failure will be treated like an implicit operation
   *  failure and behave according to `on_implicit_error`.
   */
  on_error?: NextOperation | null;
  [k: string]: unknown;
}
export interface BufferSchema {
  /**
   * If true, messages will be serialized before sending into the buffer.
   */
  serialize?: boolean | null;
  settings?: BufferSettings;
  [k: string]: unknown;
}
/**
 * Settings to describe the behavior of a buffer.
 */
export interface BufferSettings {
  retention: RetentionPolicy;
  [k: string]: unknown;
}
export interface BufferAccessSchema {
  buffers: BufferSelection;
  next: NextOperation;
  [k: string]: unknown;
}
export interface ListenSchema {
  buffers: BufferSelection;
  next: NextOperation;
  /**
   * The id of an operation that this operation is for. The id must be a `node` operation. Optional if `next` is a node operation.
   */
  target_node?: NextOperation | null;
  [k: string]: unknown;
}